fix: path validation security + writer.ts split + SDK decoupling + docs
Security (critical):
- Replace startsWith() boundary checks with path.relative() to prevent
  sibling-directory prefix attacks (e.g., /vault vs /vault-sibling)
- Fix overly-broad '..' guard that rejected legitimate filenames like ...note.md
- Upgrade readVaultFile() to use validatePathSecure() for symlink protection
- Add 13 regression tests for path boundary security

Architecture:
- Split writer.ts (1,559 lines) into 6 focused modules + barrel re-export:
  path-security, regex-safety, line-endings, markdown-structure,
  content-mutation, file-io — zero import changes in consumers
- Replace _registeredTools SDK internal access with local toolHandles map
- Bump @modelcontextprotocol/sdk dep to ^1.26.0
- Document remaining SDK internal access points for version audits

Docs & versioning:
- Add v37/v38 to ARCHITECTURE.md schema history
- Add semantic model download disclosure to README.md
- Update QUALITY_REPORT.md version references to 2.5.10
- Mark root package.json as 0.0.0-private to clarify non-semantic version
