+++ docs/QUALITY_REPORT.md
@@ -1,7 +1,7 @@
 # PROVE-IT: Graph Quality Evidence
 
 > Auto-generated by `npm run test:quality:report` on 2026-03-27 21:45:21 UTC
-> Engine version: 2.0.154
+> Engine version: 2.5.10
 
 Balanced mode currently lands at **40.2% precision / 71.7% recall / 51.5% F1** on the primary synthetic vault. Use this report for the breakdown by strictness mode, category, layer ablation, archetype, and graph topology.
 
@@ -141,4 +141,4 @@ Balanced mode currently lands at **40.2% precision / 71.7% recall / 51.5% F1** o
 - Layer ablation: each scoring layer disabled individually; F1 delta measures contribution
 - Scoring layers (13): length_filter, article_filter, exact_match, stem_match, cooccurrence, type_boost, context_boost, recency, cross_folder, hub_boost, feedback, semantic, edge_weight
 - Run: `npm run test:quality:report`
-- Engine: flywheel-memory v2.0.154
+- Engine: flywheel-memory v2.5.10
